Authenticated Volumes, TLS transport and Hostinfo Support
Authenticated volumes require a one-time password (OTP) to be entered in order
to create a volume, mount a volume, or remove a volume. If the OTP is not
entered, the volume operation is denied. This functionality is supported using
standard Docker commands and API calls. Enable authenticated volumes in a
Storage Profile or inline by configuring the volume's user to require OTP.

TLS transport volume support secures the transport used to attach the volume to
the host via TLSv1.2. Configure the volume by specifying "transport=tls" in the
volume options, inline or in Storage Profiles. The network attached (iSCSI)
volume will be secured over TLS directly to the storage target.

Hostinfo support provides metadata about the volume, pushed to the Blockbridge
management control plane. Information such as the volume options, volume type,
filesystem statistics for the volume, the host it is attached to, the Docker
containers that are using the volume, are all presented in the management
interface for the volume.