Support user scoped token for API authorization
Bump version to version 4.3.

Allow specifying the BLOCKBRIDGE_API_KEY as a user-scoped token, that
does not require system global access. If the token specified *is*
a global "system" token, then specify BLOCKBRIDGE_GLOBAL_TOKEN="1" to
change the mode of operation.

With a user scoped token, no "user" is required when creating a volume
or a profile. All volumes are and state are kept per-user. To use
different users, use different volume plugins.

Includes the following changes:

* provision with "user" api token by default
* rev version to v4.3
* volume_user not required for user_access_token
* Create new scoped authorizations only if specified token allows it
* cache lookup vol info respects when no "user" is specified
* don't lookup or use "user" in api or profile if user_access_token
* conditionally prefix log messages with a timestamp